Refactor Sidebar component: move to sidebar directory, create SidebarLink component, and update import paths.
