+++ app/search/SearchPageContent.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useState, useEffect, useCallback } from "react";
 import { useSearchParams } from "next/navigation";
-import Sidebar from "@/components/Sidebar";
+import Sidebar from "@/components/sidebar/Sidebar";
 import { CubeLoader } from "@/components/ui/CubeLoader";
 import MessageList from "@/components/search/Messages/MessageList";
 import SearchInput from "@/components/search/SearchInput";
