+++ components/sidebar/Sidebar.tsx
@@ -2,46 +2,21 @@
 import React, { useState, useEffect, useRef } from "react";
 import { motion, AnimatePresence } from "framer-motion";
 import { Box, Menu, X, Clock, Calendar, UsersRound, Home } from "lucide-react";
+import { SidebarLink } from "./SidebarLink";
 
 import Link from "next/link";
-
-interface SidebarLinkProps {
-  icon: React.ElementType;
-  label: string;
-  active?: boolean;
-  href?: string;
-}
-
-const SidebarLink: React.FC<SidebarLinkProps> = ({
-  icon: Icon,
-  label,
-  active = false,
-  href,
-}) => {
-  const content = (
-    <div
-      className={`flex items-center gap-2 px-3 py-2 rounded-xl cursor-pointer select-none transition-all duration-200 ${
-        active
-          ? "bg-white/10 text-white shadow-lg shadow-white/5"
-          : "text-white/80 hover:bg-white/5 hover:text-white hover:scale-105"
-      }`}
-    >
-      <Icon className="h-4 w-4" />
-      <span className="text-sm">{label}</span>
-    </div>
-  );
-
-  if (href) {
-    return <Link href={href}>{content}</Link>;
-  }
-
-  return content;
-};
 
 interface SidebarProps {
   open?: boolean;
   onOpenChange?: (open: boolean) => void;
 }
+
+const sidebarLinks = [
+  { icon: Home, label: "Home", href: "/" },
+  { icon: Calendar, label: "Events", href: undefined },
+  { icon: UsersRound, label: "Clubs", href: "/clubs" },
+  { icon: Clock, label: "History", href: undefined },
+];
 
 const Sidebar: React.FC<SidebarProps> = ({ open, onOpenChange }) => {
   const [internalOpen, setInternalOpen] = useState(false);
@@ -139,20 +114,15 @@ const Sidebar: React.FC<SidebarProps> = ({ open, onOpenChange }) => {
             </span>
           </Link>
           <nav className="mt-2 flex flex-col gap-1.5">
-            <SidebarLink
-              icon={Home}
-              label="Home"
-              href="/"
-              active={pathName === "/"}
-            />
-            <SidebarLink icon={Calendar} label="Events" />
-            <SidebarLink
-              icon={UsersRound}
-              label="Clubs"
-              href="/clubs"
-              active={pathName === "/clubs"}
-            />
-            <SidebarLink icon={Clock} label="History" />
+            {sidebarLinks.map((link) => (
+              <SidebarLink
+                key={link.label}
+                icon={link.icon}
+                label={link.label}
+                href={link.href}
+                pathName={pathName}
+              />
+            ))}
           </nav>
         </div>
       </motion.aside>
